@@ -15,35 +15,36 @@
     "codemirror": "git://github.com/codemirror/codemirror.git#d1d6ee2ddeaafc5d5f540b22c8d4fa58ce4bf92c",
     "core-js": "^2.4.1",
     "cross-env": "1.0.7",
-    "css-loader": "0.21.0",
+    "css-loader": "0.28.4",
     "event-emitter": "0.3.4",
     "exports-loader": "^0.6.3",
     "extend": "2.0.1",
-    "extract-text-webpack-plugin": "0.8.2",
+    "extract-text-webpack-plugin": "3.0.0",
     "file-loader": "0.9.0",
     "fixer.js": "1.4.15",
     "imports-loader": "^0.7.0",
     "in-viewport": "3.4.2",
     "jquery": "2.1.4",
+    "kotlin-runcode": "1.0.1",
     "monkberry": "4.0.8",
     "monkberry-directives": "4.0.8",
     "monkberry-events": "4.0.8",
     "monkberry-loader": "4.0.8",
     "node-sass": "3.4.1",
-    "nunjucks": "2.4.2",
-    "nunjucks-loader": "2.4.1",
+    "nunjucks": "3.0.1",
+    "nunjucks-loader": "2.4.8",
     "postcss-loader": "0.8.0",
     "query-string": "3.0.0",
     "resolve-url-loader": "1.6.0",
     "rimraf": "2.5.2",
-    "sass-loader": "3.1.1",
+    "sass-loader": "6.0.6",
     "scrolltofixed": "1.0.6",
     "smoothscroll-polyfill": "0.3.4",
     "style-loader": "0.13.0",
     "svg-fill-loader": "0.0.6",
     "url-loader": "0.5.7",
     "url-search-params": "0.6.1",
-    "webpack": "1.12.2",
+    "webpack": "^3.0.0",
     "whatwg-fetch": "2.0.2"
   },
   "devDependencies": {
